+++ pragma-sdk/tests/integration/fetchers/fetcher_configs.py
@@ -4,9 +4,10 @@ from pragma_sdk.common.fetchers.fetchers import (
     BitstampFetcher,
     CoinbaseFetcher,
     OkxFetcher,
-    StarknetAMMFetcher,
+    EkuboFetcher,
     PropellerFetcher,
     GeckoTerminalFetcher,
+    DexscreenerFetcher,
 )
 from pragma_sdk.common.fetchers.future_fetchers import (
     ByBitFutureFetcher,
@@ -255,16 +256,16 @@ ONCHAIN_FETCHER_CONFIGS = {
             ),
         ],
     },
-    "StarknetAMMFetcher": {
-        "mock_file": MOCK_DIR / "responses" / "starknet_amm.json",
-        "fetcher_class": StarknetAMMFetcher,
-        "name": "Starknet",
+    "EkuboFetcher": {
+        "mock_file": MOCK_DIR / "responses" / "ekubo.json",
+        "fetcher_class": EkuboFetcher,
+        "name": "Ekubo",
         "expected_result": [
             SpotEntry(
                 "LUSD/USD",
                 1001354537000,
                 12345,
-                "STARKNET",
+                "EKUBO",
                 PUBLISHER_NAME,
                 volume=0,
             ),
@@ -272,7 +273,30 @@ ONCHAIN_FETCHER_CONFIGS = {
                 "WBTC/USD",
                 5763275533000,
                 12345,
-                "STARKNET",
+                "EKUBO",
+                PUBLISHER_NAME,
+                volume=0,
+            ),
+        ],
+    },
+    "DexscreenerFetcher": {
+        "mock_file": MOCK_DIR / "responses" / "dexscreener.json",
+        "fetcher_class": DexscreenerFetcher,
+        "name": "Dexscreener",
+        "expected_result": [
+            SpotEntry(
+                "LUSD/USD",
+                56845000000,
+                12345,
+                "DEXSCREENER",
+                PUBLISHER_NAME,
+                volume=0,
+            ),
+            SpotEntry(
+                "WBTC/USD",
+                56845000000,
+                12345,
+                "DEXSCREENER",
                 PUBLISHER_NAME,
                 volume=0,
             ),