+++ price-pusher/price_pusher/configs/fetchers.py
@@ -13,7 +13,8 @@ from pragma_sdk.common.fetchers.fetchers import (
     OkxFetcher,
     BinanceFetcher,
     PropellerFetcher,
-    StarknetAMMFetcher,
+    EkuboFetcher,
+    DexscreenerFetcher,
 )
 from pragma_sdk.common.fetchers.future_fetchers import BinanceFutureFetcher, ByBitFutureFetcher
 
@@ -28,7 +29,8 @@ ALL_SPOT_FETCHERS: List[FetcherInterfaceT] = [
     BybitFetcher,
     BinanceFetcher,
     PropellerFetcher,
-    StarknetAMMFetcher,
+    EkuboFetcher,
+    DexscreenerFetcher,
 ]
 
 ALL_FUTURE_FETCHERS: List[FetcherInterfaceT] = [BinanceFutureFetcher, ByBitFutureFetcher]