+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## v2.0.6 (2024-08-02)
+
+### Feat
+
+- add EKUBO pairs to fetchers
+
+### Fix
+
+- pair serialization method
+
+## v2.0.5 (2024-08-02)
+
+### Feat
+
+- Dexscreener fetcher (#183)
+- Merkle Maker service (#164)
+
+### Fix
+
+- Price pusher median comparison (#182)
+
 ## v2.0.4 (2024-07-31)
 
 ### Feat